trying unifiying code for metadata and S0, S1 y S2 generation.
