+++ wrappers/json_tpl/common_activity.json
