+++ wrappers/json_tpl/common_catalog.json
@@ -1,7 +1,7 @@
 {
   "@id" : "/CATCODENAME",   
   "@type" : "Catalog",
-  "description" : "A Complete Corsika Simulation",
+  "description" : "LAGO Simulation",
   "title" : "CATCODENAME",
   "wasGeneratedBy" : "/CATCODENAME#activity",
   "_landing_page" : "https://datahub.egi.eu/dummy_hash"
