+++ wrappers/json_tpl/dataset_corsika_bin_output.json
@@ -1,17 +1,2 @@
 {
-  "@graph" : [
-    {
-     "@id" : "/CATCODENAME/FILENAME#distribution",
-     "@type" : "Distribution",
-     "accessURL" : "LANDINGPAGE",
-     "format" : "http://publications.europa.eu/resource/authority/file-type/BIN",
-     "accessService" : "/CATCODENAME/FILENAME#dataservice",
-     "mediaType" : "https://www.iana.org/assignments/media-types/application/octet-stream",
-     "publisher" :  { "@id" : "https://raw.githubusercontent.com/lagoproject/DMP/1.1/defs/lagocollaboration.jsonld" , "@type" : "lago:Organization" },
-     "creator" : { "@id" : "https://orcid.org/ORCID" , "@type" : "lago:Person" },
-     "license" : "https://creativecommons.org/licenses/by-nc-sa/4.0/",
-     "rights" : "https://raw.githubusercontent.com/lagoproject/DMP/1.1/rights/lagoCommonRights.jsonld",
-     "accessRigths" : "http://publications.europa.eu/resource/authority/access-right/RESTRICTED"
-    }
-  ]
 }
