+++ wrappers/json_tpl/dataset_corsika_lst_output.json
@@ -1,17 +1,2 @@
 {
-  "@graph" : [
-   {
-     "@id" : "/CATCODENAME/FILENAME#distribution",
-     "@type" : "Distribution",
-     "accessURL" : "LANDINGPAGE",
-     "format" : "http://publications.europa.eu/resource/authority/file-type/TXT",
-     "accessService" : "/CATCODENAME/FILENAME#dataservice",
-     "mediaType" : "http://www.iana.org/assignments/media-types/text",
-     "publisher" :  { "@id" : "https://raw.githubusercontent.com/lagoproject/DMP/1.1/defs/lagocollaboration.jsonld" , "@type" : "lago:Organization" },
-     "creator" : { "@id" : "https://orcid.org/ORCID" , "@type" : "lago:Person" },
-     "license" : "https://creativecommons.org/licenses/by-nc-sa/4.0/",
-     "rights" : "https://raw.githubusercontent.com/lagoproject/DMP/1.1/rights/lagoCommonRights.jsonld",
-     "accessRigths" : "http://publications.europa.eu/resource/authority/access-right/RESTRICTED"
-    }
-  ]
 }
